create target and struture for nef menu
