+++ project/Component/nef/Support Files/_Carbon.h
@@ -3,8 +3,5 @@
 #import <Cocoa/Cocoa.h>
 #import "NefCarbon.h"
 
-//! Project version number for NefCarbon.
 FOUNDATION_EXPORT double CarbonVersionNumber;
-
-//! Project version string for NefCarbon.
 FOUNDATION_EXPORT const unsigned char CarbonVersionString[];
