+++ project/Component/nef/Support Files/_Compiler.h
@@ -1,0 +1,7 @@
+//  Copyright © 2019 The nef Authors.
+
+#import <Cocoa/Cocoa.h>
+#import "NefCompiler.h"
+
+FOUNDATION_EXPORT double CompilerVersionNumber;
+FOUNDATION_EXPORT const unsigned char CompilerVersionString[];
