+++ project/Component/nef/Support Files/_Core.h
@@ -3,8 +3,5 @@
 #import <Cocoa/Cocoa.h>
 #import "NefCore.h"
 
-//! Project version number for Markup.
 FOUNDATION_EXPORT double CoreVersionNumber;
-
-//! Project version string for Markup.
 FOUNDATION_EXPORT const unsigned char CoreVersionString[];
