+++ project/Component/nef/Support Files/_Jekyll.h
@@ -3,8 +3,5 @@
 #import <Cocoa/Cocoa.h>
 #import "NefJekyll.h"
 
-//! Project version number for NefJekyll.
 FOUNDATION_EXPORT double JekyllVersionNumber;
-
-//! Project version string for NefJekyll.
 FOUNDATION_EXPORT const unsigned char JekyllVersionString[];
