+++ project/Component/nef/Support Files/_Markdown.h
@@ -3,8 +3,5 @@
 #import <Cocoa/Cocoa.h>
 #import "NefMarkdown.h"
 
-//! Project version number for NefMarkdown.
 FOUNDATION_EXPORT double MarkdownVersionNumber;
-
-//! Project version string for NefMarkdown.
 FOUNDATION_EXPORT const unsigned char MarkdownVersionString[];
