+++ project/Component/nef/Support Files/_Playground.h
@@ -1,0 +1,7 @@
+//  Copyright © 2019 The nef Authors.
+
+#import <Cocoa/Cocoa.h>
+#import "NefPlayground.h"
+
+FOUNDATION_EXPORT double PlaygroundVersionNumber;
+FOUNDATION_EXPORT const unsigned char PlaygroundVersionString[];
