+++ project/Component/nef/Support Files/_SwiftPlayground.h
@@ -3,8 +3,5 @@
 #import <Cocoa/Cocoa.h>
 #import "NefSwiftPlayground.h"
 
-//! Project version number for NefCarbon.
 FOUNDATION_EXPORT double SwiftPlaygroundAPIVersionNumber;
-
-//! Project version string for NefCarbon.
 FOUNDATION_EXPORT const unsigned char SwiftPlaygroundAPIVersionString[];
