+++ project/Component/nef/Support Files/nef.h
@@ -1,11 +1,8 @@
 //  Copyright © 2019 The nef Authors.
 
 #import <Cocoa/Cocoa.h>
-#import "NefModels.h"
 #import "NefCommon.h"
+#import "NefModels.h"
 
-//! Project version number for Nef.
 FOUNDATION_EXPORT double NefVersionNumber;
-
-//! Project version string for Nef.
 FOUNDATION_EXPORT const unsigned char NefVersionString[];
