+++ project/UI/Compiler/Compiler.swift
@@ -1,0 +1,43 @@
+//  Copyright © 2020 The nef Authors.
+
+import Foundation
+import CLIKit
+import nef
+import Bow
+import BowEffects
+
+enum CompilerCommands: String {
+    case project
+    case cached = "use-cache"
+}
+
+
+@discardableResult
+public func compiler() -> Either<CLIKit.Console.Error, Void> {
+    let console = Console(script: "nefc",
+                                  description: "Compile Xcode Playground",
+                                  arguments: .init(name: CompilerCommands.project.rawValue, placeholder: "path-to-input", description: "path to the folder containing Xcode Playgrounds to render"),
+                                             .init(name: CompilerCommands.cached.rawValue, placeholder: "", description: "use cached dependencies if it is possible.", isFlag: true, default: "false"))
+
+
+    func arguments(console: CLIKit.Console) -> IO<CLIKit.Console.Error, (input: URL, cached: Bool)> {
+        console.input().flatMap { args in
+            guard let inputPath = args[CompilerCommands.project.rawValue]?.trimmingEmptyCharacters.expandingTildeInPath,
+                  let cached = Bool(args[CompilerCommands.cached.rawValue] ?? "") else { return IO.raiseError(.arguments) }
+            
+            return IO.pure((input: URL(fileURLWithPath: inputPath, isDirectory: true), cached: cached))
+        }^
+    }
+    
+    return arguments(console: console)
+        .flatMap { (input, cached) in
+            nef.Compiler.compile(nefPlayground: input, cached: cached)
+               .provide(console)^
+               .mapError { _ in .render() }
+               .foldM({ e in console.exit(failure: "rendering Xcode Playgrounds from '\(input.path)'. \(e)") },
+                      { _ in console.exit(success: "'\(input.path)' compiled successfully")            })    }^
+        .reportStatus(in: console)
+        .foldM({ e in console.exit(failure: "\(e)")        },
+               { success in console.exit(success: success) })
+        .unsafeRunSyncEither()
+}
