+++ project/UI/Nef/main.swift
@@ -1,0 +1,20 @@
+//  Copyright © 2020 The nef Authors.
+
+import Foundation
+import CLIKit
+import nef
+import Bow
+import BowEffects
+
+enum NefCommand {
+    
+}
+
+
+@discardableResult
+public func main() -> Either<CLIKit.Console.Error, Void> {
+    fatalError()
+}
+
+// #: - MAIN <launcher>
+main()
